Reorganize SIMD algorithm.hpp and inner_product.hpp
