+++ src/QMCWaveFunctions/Fermion/DiracMatrixComputeOMPTarget.hpp
@@ -21,7 +21,7 @@
 #include "type_traits/complex_help.hpp"
 #include "type_traits/template_types.hpp"
 #include "Concurrency/OpenMP.h"
-#include "CPU/SIMD/simd.hpp"
+#include "CPU/SIMD/algorithm.hpp"
 #include "ResourceCollection.h"
 
 namespace qmcplusplus
